+++ lib/mutant.rb
@@ -135,7 +135,6 @@ require 'mutant/mutator/node/send'
 require 'mutant/mutator/node/send/binary'
 require 'mutant/mutator/node/send/conditional'
 require 'mutant/mutator/node/send/attribute_assignment'
-require 'mutant/mutator/node/send/index'
 require 'mutant/mutator/node/when'
 require 'mutant/mutator/node/class'
 require 'mutant/mutator/node/define'
@@ -151,6 +150,8 @@ require 'mutant/mutator/node/regopt'
 require 'mutant/mutator/node/resbody'
 require 'mutant/mutator/node/rescue'
 require 'mutant/mutator/node/match_current_line'
+require 'mutant/mutator/node/index'
+require 'mutant/mutator/node/procarg_zero'
 require 'mutant/loader'
 require 'mutant/context'
 require 'mutant/scope'
