+++ lib/mutant/ast/types.rb
@@ -182,10 +182,12 @@ module Mutant
       # Nodes that are NOT generated by parser but used by mutant / unparser.
       GENERATED = Set.new(%i[empty]).freeze
 
-      EXTRA = Set.new(GENERATED + REGEXP)
+      # Nodes missing from parser metadata
+      MISSING = Set.new(%i[index indexasgn lambda procarg0]).freeze
 
-      # All node types mutant handles
-      ALL = Set.new((Parser::Meta::NODE_TYPES + EXTRA) - BLACKLIST).freeze
+      ALL = Set.new(
+        (Parser::Meta::NODE_TYPES + GENERATED + REGEXP + MISSING) - BLACKLIST
+      ).freeze
     end # Types
   end # AST
 end # Mutant
