+++ lib/mutant/meta/example/dsl.rb
@@ -99,7 +99,7 @@ module Mutant
         def node(input)
           case input
           when String
-            Unparser::Preprocessor.run(::Parser::CurrentRuby.parse(input))
+            Unparser::Preprocessor.run(Unparser.parse(input))
           when ::Parser::AST::Node
             input
           else
