+++ lib/mutant/mutator/node/arguments.rb
@@ -24,9 +24,21 @@ module Mutant
         # @return [undefined]
         def emit_argument_presence
           emit_type
+
           Util::Array::Presence.call(children).each do |children|
-            emit_type(*children)
+            if children.one? && n_mlhs?(Mutant::Util.one(children))
+              emit_procarg(Mutant::Util.one(children))
+            else
+              emit_type(*children)
+            end
           end
+        end
+
+        # Emit procarg form
+        #
+        # @return [undefined]
+        def emit_procarg(arg)
+          emit_type(s(:procarg0, *arg))
         end
 
         # Emit argument mutations
