+++ lib/mutant/mutator/node/block.rb
@@ -17,7 +17,7 @@ module Mutant
         # @return [undefined]
         def dispatch
           emit_singletons
-          emit(send)
+          emit(send) unless n_lambda?(send)
           emit_send_mutations(&method(:n_send?))
           emit_arguments_mutations
