+++ lib/mutant/mutator/node/noop.rb
@@ -7,7 +7,7 @@ module Mutant
       # Mutation emitter to handle noop nodes
       class Noop < self
 
-        handle(:__ENCODING__, :block_pass, :cbase)
+        handle(:__ENCODING__, :block_pass, :cbase, :lambda)
 
       private
