+++ lib/mutant/zombifier.rb
@@ -115,7 +115,7 @@ module Mutant
     #
     # @return [Parser::AST::Node]
     def namespaced_node(source_path)
-      s(:module, s(:const, nil, namespace), ::Parser::CurrentRuby.parse(source_path.read))
+      s(:module, s(:const, nil, namespace), Unparser.parse(source_path.read))
     end
 
   end # Zombifier
