+++ meta/lambda.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+Mutant::Meta::Example.add :block, :lambda do
+  source '->() {}'
+
+  singleton_mutations
+
+  mutation '->() { raise }'
+end
