+++ mutant.gemspec
@@ -36,7 +36,7 @@ Gem::Specification.new do |gem|
   gem.add_runtime_dependency('parser',        '~> 2.5.1')
   gem.add_runtime_dependency('procto',        '~> 0.0.2')
   gem.add_runtime_dependency('regexp_parser', '~> 1.2')
-  gem.add_runtime_dependency('unparser',      '~> 0.3.0')
+  gem.add_runtime_dependency('unparser',      '~> 0.4.0')
 
   gem.add_development_dependency('devtools', '~> 0.1.22')
   gem.add_development_dependency('parallel', '~> 1.3')
