+++ spec/unit/mutant/ast/meta/send/receiver_possible_top_level_const_predicate_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Mutant::AST::Meta::Send, '#receiver_possible_top_level_const?' do
   subject { described_class.new(node).receiver_possible_top_level_const? }
 
   def parse(source)
-    Parser::CurrentRuby.parse(source)
+    Unparser.parse(source)
   end
 
   context 'when implicit top level const' do
